Introduce ToBeFixedForIsolatedProjects annotation
